app/backend/requirements.txt
fastapi>=0.115.0,<1.0
starlette>=0.46.2,<1.0
uvicorn[standard]>=0.34.0,<1.0
websockets>=15.0.1,<16.0.0
httpx>=0.28.1,<1.0
requests>=2.32.4,<3.0
pydantic>=2.11.0,<3.0
pydantic-core>=2.20.0,<3.0
pydantic-settings>=2.5.2,<3.0
python-dotenv==1.0.0
google-generativeai>=0.5.2
google-ai-generativelanguage>=0.6.2
google-adk>=1.17.0
census==0.8.19
us==3.1.1
shapely==2.1.2
setuptools>=68.0
